fix code

+++ BOJ_4153/MainTest.java
@@ -1,6 +1,5 @@
 package BOJ_4153;
 
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
@@ -8,19 +7,19 @@ class MainTest {
 
 	@Test
 	void test1() {
-		int[] arr = {6,8,10};
+		long[] arr = {6,8,10};
 		System.out.println(Main.solution(arr));
 	}
 	
 	@Test
 	void test2() {
-		int[] arr = {25,52,60};
+		long[] arr = {25,52,60};
 		System.out.println(Main.solution(arr));
 	}
 
 	@Test
 	void test3() {
-		int[] arr = {5,12,13};
+		long[] arr = {5,12,13};
 		System.out.println(Main.solution(arr));
 	}
